docker_run/requirements.txt
pandas==0.24.1
scipy==1.2.1
numpy==1.16.1
PyYAML==3.13
